changed return type from float to double

+++ framework/shape.hpp
@@ -5,7 +5,7 @@
 class Shape
 {
 	public:
-		virtual float area() const = 0;
-		virtual float volume() const = 0;
+		virtual double area() const = 0;
+		virtual double volume() const = 0;
 };
 #endif
